The preliminary version of lstm
In this version, LSTM is implemented for the case when there is 1 cell in each layer, and the size of the input sequence of vectors is equal to the size of the output sequence of vectors. The project is compiled. The test data in the experiment project is set correctly, but there is no test for training